+++ lib/src/analyzers/lint_analyzer/anti_patterns/patterns_factory.dart
@@ -5,15 +5,15 @@ import 'models/pattern.dart';
 
 typedef CreatePattern = Pattern Function(
   Map<String, Object> patternSettings,
-  Map<String, Object> metricstTresholds,
+  Map<String, Object> metricsThresholds,
 );
 
 final _implementedPatterns = <String, CreatePattern>{
-  LongMethod.patternId: (settings, tresholds) =>
-      LongMethod(patternSettings: settings, metricstTresholds: tresholds),
-  LongParameterList.patternId: (settings, tresholds) => LongParameterList(
+  LongMethod.patternId: (settings, thresholds) =>
+      LongMethod(patternSettings: settings, metricsThresholds: thresholds),
+  LongParameterList.patternId: (settings, thresholds) => LongParameterList(
         patternSettings: settings,
-        metricstTresholds: tresholds,
+        metricsThresholds: thresholds,
       ),
 };