+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/halstead_volume/halstead_volume_ast_visitor.dart
@@ -9,16 +9,16 @@ class HalsteadVolumeAstVisitor extends RecursiveAstVisitor<void> {
   final _operators = <String, int>{};
   final _operands = <String, int>{};
 
-  /// the number of operators
+  /// The number of operators.
   int get operators => _operators.values.sum;
 
-  /// the number of unique operators
+  /// The number of unique operators.
   int get uniqueOperators => _operators.keys.length;
 
-  /// the number of operands
+  /// The number of operands.
   int get operands => _operands.values.sum;
 
-  /// the number of unique operands
+  /// The number of unique operands.
   int get uniqueOperands => _operands.keys.length;
 
   @override