+++ lib/src/analyzers/lint_analyzer/reporters/lint_report_params.dart
@@ -1,4 +1,4 @@
-/// Represents additional lint reporter params
+/// Represents additional lint reporter params.
 class LintReportParams {
   final bool congratulate;