+++ lib/src/analyzers/unused_code_analyzer/reporters/unused_code_report_params.dart
@@ -1,4 +1,4 @@
-/// Represents additional unused code reporter params
+/// Represents additional unused code reporter params.
 class UnusedCodeReportParams {
   final bool congratulate;