+++ lib/src/analyzers/unused_l10n_analyzer/models/unused_l10n_issue.dart
@@ -2,7 +2,7 @@ import 'package:source_span/source_span.dart';
 
 /// Represents an issue detected by the unused localization check.
 class UnusedL10nIssue {
-  /// The name of a class member, which is unused
+  /// The name of a class member, which is unused.
   final String memberName;
 
   /// The source location associated with this issue.