+++ lib/src/reporters/models/json_reporter.dart
@@ -5,7 +5,7 @@ import 'package:meta/meta.dart';
 import 'file_report.dart';
 import 'reporter.dart';
 
-/// Machine-readable report in JSON format
+/// Machine-readable report in JSON format.
 abstract class JsonReporter<T extends FileReport, S, P>
     extends Reporter<T, S, P> {
   static const String id = 'json';