+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* **Breaking Change:** cli arguments `--fatal-unused` and `--fatal-warnings` activate by default.
+* chore: restrict `analyzer` version to `>=2.8.0 <3.3.0`.
+
 ## 4.11.0
 
 * feat: add static code diagnostics `format-comment`.